edit migrations

+++ giftstore_website/database/migrations/2022_06_07_143624_create_tbl_bill_order.php
@@ -20,7 +20,7 @@ class CreateTblBillOrder extends Migration
             $table->integer('id_stock');//foreign key
             $table->datetime('date_order');
             $table->integer('quantity');
-            $table->float('total_price');
+            $table->bigInteger('total_price');
             $table->string('status');
             $table->timestamps();
         });
